ecomm_app/requirements.txt
streamlit
Pillow
boto3
botocore
